avoid inappropriate use of long
git-svn-id: https://svn.r-project.org/R/trunk@64677 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/arithmetic.c
@@ -1426,7 +1426,7 @@ static SEXP math2B(SEXP sa, SEXP sb, double (*f)(double, double, double *),
     double ai, bi, *a, *b, *y;
     int naflag;
     double amax, *work;
-    long nw;
+    size_t nw;
 
     if (!isNumeric(sa) || !isNumeric(sb))
 	errorcall(lcall, R_MSG_NONNUM_MATH);
@@ -1443,8 +1443,8 @@ static SEXP math2B(SEXP sa, SEXP sb, double (*f)(double, double, double *),
 	if (av > amax) amax = av;
     }
     const void *vmax = vmaxget();
-    nw = 1 + (long)floor(amax);
-    work = (double *) R_alloc((size_t) nw, sizeof(double));
+    nw = 1 + (size_t)floor(amax);
+    work = (double *) R_alloc(nw, sizeof(double));
 
     mod_iterate(na, nb, ia, ib) {
 //	if ((i+1) % NINTERRUPT == 0) R_CheckUserInterrupt();
@@ -1761,7 +1761,7 @@ static SEXP math3B(SEXP sa, SEXP sb, SEXP sc,
     double ai, bi, ci, *a, *b, *c, *y;
     int naflag;
     double amax, *work;
-    long nw;
+    size_t nw;
 
     SETUP_Math3;
 
@@ -1773,8 +1773,8 @@ static SEXP math3B(SEXP sa, SEXP sb, SEXP sc,
 	if (av > amax) amax = av;
     }
     const void *vmax = vmaxget();
-    nw = 1 + (long)floor(amax);
-    work = (double *) R_alloc((size_t) nw, sizeof(double));
+    nw = 1 + (size_t)floor(amax);
+    work = (double *) R_alloc(nw, sizeof(double));
 
     mod_iterate3 (na, nb, nc, ia, ib, ic) {
 //	if ((i+1) % NINTERRUPT == 0) R_CheckUserInterrupt();
