+++ src/main/eval.c
@@ -101,7 +101,7 @@ static int R_Profiling = 0;
 
 static FILE *R_ProfileOutfile = NULL;
 static int R_Mem_Profiling=0;
-extern void get_current_mem(unsigned long *,unsigned long *,unsigned long *); /* in memory.c */
+extern void get_current_mem(size_t *,size_t *,size_t *); /* in memory.c */
 extern unsigned long get_duplicate_counter(void);  /* in duplicate.c */
 extern void reset_duplicate_counter(void);         /* in duplicate.c */
 static int R_GC_Profiling = 0;                     /* indicates GC profiling */
@@ -199,7 +199,7 @@ static void doprof(int sig)  /* sig is ignored in Windows */
 {
     RCNTXT *cptr;
     char buf[PROFBUFSIZ];
-    unsigned long bigv, smallv, nodes;
+    size_t bigv, smallv, nodes;
     size_t len;
     int prevnum = R_Line_Profiling;
